move hostlist to zsutand Fixes #32
