+++ InvenTree/web/src/src/defaults.tsx
@@ -1,0 +1,7 @@
+import { HostList } from './states';
+
+export const defaultlist: HostList = {
+  "https://demo.inventree.org": { host: "https://demo.inventree.org/api/", name: "InvenTree Demo" },
+  "https://sample.app.invenhost.com": { host: "https://sample.app.invenhost.com/api/", name: "InvenHost: Sample" },
+  "http://localhost:8000": { host: "http://localhost:8000/api/", name: "Localhost" },
+};
