+++ InvenTree/web/src/src/states.tsx
@@ -1,11 +1,23 @@
 import create from 'zustand'
 import { persist } from 'zustand/middleware'
 
+// helpers
+export interface HostProps {
+    host: string,
+    name: string,
+}
+
+export interface HostList {
+    [key: string]: HostProps
+}
+
+// states
 interface SesstionSettings {
     autoupdate: boolean,
     toggleAutoupdate: () => void,
     host: string,
     hostKey: string,
+    hostList: HostList,
     setHost: (newHost: string, newHostKey: string) => void,
 }
 
@@ -17,6 +29,7 @@ export const useSessionSettings = create<SesstionSettings>(
             toggleAutoupdate: () => set((state) => ({ autoupdate: !state.autoupdate })),
             host: '',
             hostKey: '',
+            hostList: {},
             setHost: (newHost, newHostKey) => set({ host: newHost, hostKey: newHostKey }),
         }),
         {
